# docs/storage-api.openapi.yaml
openapi: 3.1.0
info:
  title: Deniz Cloud Storage API
  version: 1.0.0
  description: |
    Cloud storage API for Deniz Cloud self-hosted home server.
    Provides file management (upload, download, rename, move, delete),
    folder CRUD, and resumable uploads via the TUS protocol.

    **Base URL:** `https://storage.denizlg24.com`

    ## Authentication

    All endpoints under `/api/uploads/*`, `/api/folders/*`, and `/api/files/*` require authentication.
    Auth endpoints (`/api/auth/me`, `/api/auth/logout`) also require authentication; login is public.

    Two authentication methods are supported:

    - **Bearer Token (JWT):** Obtained via `POST /api/auth/login`. Pass as `Authorization: Bearer <token>`.
    - **API Key:** Pass as `X-API-Key: <key>` header.

    ## Access Control

    - **Private folders** (`/{userId}/...`): Only the owning user can read/write.
    - **Shared folders** (`/shared/...`): All authenticated users can read. Only the owner or a superuser can modify/delete.

    ## Storage Tiering

    Files are stored on SSD (hot) or HDD (cold) based on disk usage watermarks.
    SSD files preserve directory structure; HDD files use a flat layout keyed by file ID.

    ## TUS Resumable Uploads

    File uploads use the [TUS protocol](https://tus.io/) v1.0.0 with `creation` and `termination` extensions.
    Upload sessions expire after 24 hours. Expired sessions are cleaned up automatically.

    ## Error Format

    All errors follow a consistent shape:
...
    {
...
servers:
...
paths:
...
  /api/folders/roots:
    get:
      operationId: getRootFolders
      summary: Get root folders
      description: |
        Returns the authenticated user's personal root folder and the shared root folder.
        Both are lazily created if they do not already exist.

        - **User root:** `/{userId}` — private to the user
        - **Shared root:** `/shared` — accessible to all authenticated users
      tags: [Folders]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      responses:
        "200":
          description: Root folders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_RootFolders"
              example:
                data:
                  userRoot:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    path: /550e8400-e29b-41d4-a716-446655440000
                    name: 550e8400-e29b-41d4-a716-446655440000
                  sharedRoot:
                    id: 660e8400-e29b-41d4-a716-446655440000
                    path: /shared
                    name: shared
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/folders:
    post:
      operationId: createFolder
      summary: Create a folder
      description: |
        Creates a new folder under the specified parent. The folder name is normalized
        to snake_case. A corresponding directory is created on disk (SSD storage).

        **Name normalization:**
        - Converted to snake_case
        - Invalid characters (path separators, null bytes, etc.) are rejected
      tags: [Folders]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFolderRequest"
            example:
              name: My Documents
              parentId: 550e8400-e29b-41d4-a716-446655440000
      responses:
        "201":
          description: Folder created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_FolderCreated"
              example:
                data:
                  id: 770e8400-e29b-41d4-a716-446655440000
                  path: /550e8400-e29b-41d4-a716-446655440000/my_documents
                  name: my_documents
                  parentId: 550e8400-e29b-41d4-a716-446655440000
                  createdAt: "2026-03-13T12:00:00.000Z"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingName:
                  summary: Missing folder name
                  value:
                    error:
                      code: MISSING_NAME
                      message: Folder name is required
                missingParentId:
                  summary: Missing parentId
                  value:
                    error:
                      code: MISSING_PARENT_ID
                      message: parentId is required
                invalidName:
                  summary: Invalid folder name
                  value:
                    error:
                      code: INVALID_NAME
                      message: Name contains invalid characters
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          description: Parent folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: PARENT_NOT_FOUND
                  message: Parent folder not found
        "409":
          description: Folder already exists at target path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FOLDER_EXISTS
                  message: A folder already exists at this path
  /api/folders/{id}:
    get:
      operationId: getFolder
      summary: Get folder details
      description: Returns metadata for a single folder.
      tags: [Folders]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FolderId"
      responses:
        "200":
          description: Folder details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_FolderDetail"
              example:
                data:
                  id: 770e8400-e29b-41d4-a716-446655440000
                  path: /550e8400-e29b-41d4-a716-446655440000/documents
                  name: documents
                  parentId: 550e8400-e29b-41d4-a716-446655440000
                  ownerId: 550e8400-e29b-41d4-a716-446655440000
                  createdAt: "2026-03-01T00:00:00.000Z"
                  updatedAt: "2026-03-01T00:00:00.000Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/FolderNotFound"

    patch:
      operationId: updateFolder
      summary: Rename or move a folder
      description: |
        Renames and/or moves a folder. At least one of `name` or `parentId` must be provided.
        Root folders (`/{userId}` and `/shared`) cannot be modified.

        When a folder is moved, all descendant folders and files have their paths updated
        atomically within a database transaction. The corresponding directory is moved on disk.

        **Validation:**
        - Cannot move a folder into itself or any of its descendants (circular move)
        - Target path must not already be occupied by another folder
        - User must have access to both the folder and the target parent
      tags: [Folders]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FolderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFolderRequest"
            examples:
              rename:
                summary: Rename folder
                value:
                  name: new_name
              move:
                summary: Move to different parent
                value:
                  parentId: 880e8400-e29b-41d4-a716-446655440000
              both:
                summary: Rename and move
                value:
                  name: new_name
                  parentId: 880e8400-e29b-41d4-a716-446655440000
      responses:
        "200":
          description: Folder updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_FolderUpdated"
              example:
                data:
                  id: 770e8400-e29b-41d4-a716-446655440000
                  path: /550e8400-e29b-41d4-a716-446655440000/new_name
                  name: new_name
                  parentId: 550e8400-e29b-41d4-a716-446655440000
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                nothingToUpdate:
                  summary: No changes provided
                  value:
                    error:
                      code: NOTHING_TO_UPDATE
                      message: Provide name or parentId
                circularMove:
                  summary: Circular move detected
                  value:
                    error:
                      code: CIRCULAR_MOVE
                      message: Cannot move a folder into itself or its descendant
                invalidName:
                  summary: Invalid folder name
                  value:
                    error:
                      code: INVALID_NAME
                      message: Name contains invalid characters
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Access denied or root folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                cannotModifyRoot:
                  summary: Cannot modify root folders
                  value:
                    error:
                      code: CANNOT_MODIFY_ROOT
                      message: Cannot rename or move root folders
                accessDenied:
                  summary: Not the owner
                  value:
                    error:
                      code: ACCESS_DENIED
                      message: You do not have access to this folder
        "404":
          description: Folder or target parent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                folderNotFound:
                  summary: Folder not found
                  value:
                    error:
                      code: FOLDER_NOT_FOUND
                      message: Folder not found
                parentNotFound:
                  summary: Target parent not found
                  value:
                    error:
                      code: PARENT_NOT_FOUND
                      message: Target parent folder not found
        "409":
          description: Folder already exists at target path
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FOLDER_EXISTS
                  message: A folder already exists at the target path

    delete:
      operationId: deleteFolder
      summary: Delete a folder
      description: |
        Deletes an empty folder. The folder must contain no subfolders or files.
        Root folders (`/{userId}` and `/shared`) cannot be deleted.

        The directory is removed from disk and the database record is deleted.
      tags: [Folders]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FolderId"
      responses:
        "200":
          description: Folder deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse_DeletedId"
              example:
                data:
                  id: 770e8400-e29b-41d4-a716-446655440000
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Access denied or root folder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                cannotDeleteRoot:
                  summary: Cannot delete root folders
                  value:
                    error:
                      code: CANNOT_DELETE_ROOT
                      message: Cannot delete root folders
                accessDenied:
                  summary: Not the owner
                  value:
                    error:
                      code: ACCESS_DENIED
                      message: You do not have access to this folder
        "404":
          $ref: "#/components/responses/FolderNotFound"
        "409":
          description: Folder is not empty
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FOLDER_NOT_EMPTY
                  message: Folder is not empty. Delete all contents first.
  /api/folders/{id}/contents:
    get:
      operationId: listFolderContents
      summary: List folder contents
      description: |
        Returns subfolders and files within a folder. Subfolders are returned in full
        (ordered by name); files are paginated (ordered by `createdAt` descending).

        The `pagination` object applies to files only. Subfolders are always returned
        in their entirety.
      tags: [Folders]
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/FolderId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Folder contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedResponse_FolderContents"
              example:
                data:
                  folder:
                    id: 770e8400-e29b-41d4-a716-446655440000
                    path: /550e8400-e29b-41d4-a716-446655440000/documents
                    name: documents
                    parentId: 550e8400-e29b-41d4-a716-446655440000
                  subfolders:
                    - id: 880e8400-e29b-41d4-a716-446655440000
                      name: reports
                      path: /550e8400-e29b-41d4-a716-446655440000/documents/reports
                      parentId: 770e8400-e29b-41d4-a716-446655440000
                      createdAt: "2026-03-01T00:00:00.000Z"
                  files:
                    - id: 990e8400-e29b-41d4-a716-446655440000
                      filename: report.pdf
                      path: /550e8400-e29b-41d4-a716-446655440000/documents/report.pdf
                      mimeType: application/pdf
                      sizeBytes: 2048000
                      tier: ssd
                      createdAt: "2026-03-01T00:00:00.000Z"
                      updatedAt: "2026-03-01T00:00:00.000Z"
                pagination:
                  page: 1
                  limit: 50
                  total: 1
                  totalPages: 1
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccessDenied"
        "404":
          $ref: "#/components/responses/FolderNotFound"

  # ──────────────────────────────────────────────
  # Files
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Session JWT obtained via `POST /api/auth/login`.
        Tokens expire after 24 hours by default.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key for programmatic access. Created via the admin panel.
        Pass the raw key value (not hashed).
  parameters:
...
    FolderId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Folder UUID
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (minimum 1)
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Items per page (1-100, default 50)
  responses:
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            noAuth:
              summary: No authentication provided
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required
            tokenInvalid:
              summary: JWT is invalid or expired
              value:
                error:
                  code: TOKEN_INVALID
                  message: Invalid or expired token
            sessionInvalid:
              summary: Session revoked or expired in DB
              value:
                error:
                  code: SESSION_INVALID
                  message: Session expired or revoked
            apiKeyInvalid:
              summary: API key not found
              value:
                error:
                  code: INVALID_API_KEY
                  message: Invalid API key
            apiKeyExpired:
              summary: API key past expiry
              value:
                error:
                  code: API_KEY_EXPIRED
                  message: API key expired
    AccessDenied:
      description: Access denied to the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: ACCESS_DENIED
              message: You do not have access to this folder
    FolderNotFound:
      description: Folder not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FOLDER_NOT_FOUND
              message: Folder not found
...
  schemas:
...
    CreateFolderRequest:
      type: object
      required: [name, parentId]
      properties:
        name:
          type: string
          description: |
            Folder name. Will be normalized to snake_case.
            Must not contain path separators, null bytes, or other invalid characters.
        parentId:
          type: string
          format: uuid
          description: UUID of the parent folder
    UpdateFolderRequest:
      type: object
      properties:
        name:
          type: string
          description: New folder name (will be normalized to snake_case)
        parentId:
          type: string
          format: uuid
          description: UUID of the new parent folder (to move the folder)
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
...
    RootFolder:
      type: object
      required: [id, path, name]
      properties:
        id:
          type: string
          format: uuid
        path:
          type: string
          description: Virtual path (e.g., `/{userId}` or `/shared`)
        name:
          type: string
...
    ApiResponse_RootFolders:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [userRoot, sharedRoot]
          properties:
            userRoot:
              $ref: "#/components/schemas/RootFolder"
            sharedRoot:
              $ref: "#/components/schemas/RootFolder"
    ApiResponse_FolderCreated:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/FolderSummary"
    ApiResponse_FolderDetail:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/FolderDetail"
    ApiResponse_FolderUpdated:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id, path, name]
          properties:
            id:
              type: string
              format: uuid
            path:
              type: string
            name:
              type: string
            parentId:
              type: ["string", "null"]
              format: uuid
...
    ApiResponse_DeletedId:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [id]
          properties:
            id:
              type: string
              format: uuid
    PaginatedResponse_FolderContents:
      type: object
      required: [data, pagination]
      properties:
        data:
          $ref: "#/components/schemas/FolderContents"
        pagination:
          $ref: "#/components/schemas/Pagination"
...